chore: move components to features
